requirements.txt
absl-py==0.2.0
adafruit-io==2.0.17
astor==0.6.2
Augmentor==0.2.2
backports.weakref==1.0.post1
beautifulsoup4==4.6.3
bleach==1.5.0
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.11.0
gTTS==2.0.1
gTTS-token==1.1.1
html5lib==0.9999999
idna==2.7
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
paho-mqtt==1.4.0
pandas==0.22.0
pbr==4.2.0
Pillow==5.2.0
protobuf==3.5.2.post1
pyserial==3.4
python-dateutil==2.7.2
pytz==2018.5
requests==2.19.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tqdm==4.25.0
urllib3==1.23
Werkzeug==0.14.1
